+++ container_service_extension/def_/utils.py
@@ -4,8 +4,6 @@
 from enum import Enum
 from enum import unique
 
-from container_service_extension import shared_constants as shared_constants
-from container_service_extension.cloudapi import constants as cloudapi_constants  # noqa: E501
 from container_service_extension.cloudapi.cloudapi_client import CloudApiClient
 import container_service_extension.exceptions as excptn
 
@@ -134,88 +132,3 @@ def generate_entity_type_id(vendor, nss, version):
     :rtype str
     """
     return f"{DEF_ENTITY_TYPE_ID_PREFIX}:{vendor}:{nss}:{version}"
-
-
-def get_all_def_ent_acl(cloudapi_client, de_id):
-    """Get all def entity acl values from all pages.
-
-    :param str de_id: defined entity id
-
-    :return dict of user id keys and a dictionary values containing the
-        acl entry id and access level id
-    """
-    rel_url_path = f'entities/{de_id}/accessControls'
-    user_acl_info = {}
-    curr_page, page_cnt = 0, 1
-    while curr_page < page_cnt:
-        query_str = f'?{shared_constants.PAGE}={curr_page + 1}' \
-                    f'&{shared_constants.PAGE_SIZE}=' \
-                    f'{shared_constants.DEFAULT_PAGE_SZ}'
-        de_acl_response = cloudapi_client.do_request(
-            method=shared_constants.RequestMethod.GET,
-            cloudapi_version=cloudapi_constants.CLOUDAPI_VERSION_1_0_0,
-            resource_url_relative_path=(rel_url_path + query_str))
-        for acl_entry in de_acl_response.get('values'):
-            user_id = acl_entry[shared_constants.AccessControlKey.MEMBER_ID]  # noqa: E501
-            user_acl_info[user_id] = {
-                shared_constants.AccessControlKey.ID:
-                    acl_entry[shared_constants.AccessControlKey.ID],
-                shared_constants.AccessControlKey.ACCESS_LEVEL_ID:
-                    acl_entry[shared_constants.AccessControlKey.ACCESS_LEVEL_ID]  # noqa: E501
-            }
-        curr_page = int(de_acl_response.get('page', 1))
-        page_cnt = int(de_acl_response.get('pageCount', 1))
-    return user_acl_info
-
-
-def update_native_def_entity_acl(cloudapi_client, de_id, update_acl_entries,
-                                 prev_user_acl_info):
-    """Update native defined entity acl.
-
-    :param CloudApiClient cloudapi_client: cloudapi vlient
-    :param str de_id: defined entity id
-    :param list update_acl_entries: list of dict entries containing the
-        'memberId' and 'accessLevelId' fields
-    :param list prev_user_acl_info: dict mapping user id to dict of
-        acl entry id and acl level id
-
-    :return: dictionary of memberId keys and access level values
-    """
-    own_prev_user_acl_info = prev_user_acl_info.copy()
-
-    # Share defined entity
-    user_acl_level_dict = {}
-    access_controls_path = f'entities/{de_id}/accessControls'
-    payload = {
-        shared_constants.AccessControlKey.GRANT_TYPE:
-            shared_constants.MEMBERSHIP_GRANT_TYPE,
-        shared_constants.AccessControlKey.MEMBER_ID: None,
-        shared_constants.AccessControlKey.ACCESS_LEVEL_ID: None
-    }
-    for acl_entry in update_acl_entries:
-        user_id = acl_entry[shared_constants.AccessControlKey.MEMBER_ID]
-        acl_level = acl_entry[
-            shared_constants.AccessControlKey.ACCESS_LEVEL_ID]
-        payload[shared_constants.AccessControlKey.MEMBER_ID] = user_id
-        payload[shared_constants.AccessControlKey.ACCESS_LEVEL_ID] = acl_level
-        user_acl_level_dict[user_id] = acl_level
-        cloudapi_client.do_request(
-            method=shared_constants.RequestMethod.POST,
-            cloudapi_version=cloudapi_constants.CLOUDAPI_VERSION_1_0_0,
-            resource_url_relative_path=access_controls_path,
-            payload=payload)
-
-        # Remove entry from previous user acl info
-        if own_prev_user_acl_info.get(user_id):
-            del own_prev_user_acl_info[user_id]
-
-    # Delete def entity acl entries not in update_acl_entries
-    for _, acl_info in own_prev_user_acl_info.items():
-        delete_path = access_controls_path + \
-            f'/{acl_info[shared_constants.AccessControlKey.ID]}'
-        cloudapi_client.do_request(
-            method=shared_constants.RequestMethod.DELETE,
-            cloudapi_version=cloudapi_constants.CLOUDAPI_VERSION_1_0_0,
-            resource_url_relative_path=delete_path)
-
-    return user_acl_level_dict
